+++ clang-tools-extra/clang-tidy/ClangTidyForceLinker.h
@@ -94,6 +94,10 @@ extern volatile int MiscModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED MiscModuleAnchorDestination =
     MiscModuleAnchorSource;
 
+extern volatile int MisracModuleAnchorSource;
+static int LLVM_ATTRIBUTE_UNUSED MisracModuleAnchorDestination =
+    MisracModuleAnchorSource;
+
 // This anchor is used to force the linker to link the ModernizeModule.
 extern volatile int ModernizeModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED ModernizeModuleAnchorDestination =